moved MoleculesPanel into subfolder and moved some methods to Utilities file
